Fixes pluck bug
There was a bug that pluck couldn't be called on nil. The reason for this was that find_by_user_id_and_skill_id returns a ruby object and not a relational object. If someone is reading this, congratulations, you have found a true easter-egg. I wish I could give you something, but all I can say is that it is almost midnight on a friday and I have been trying to implement this feature for all of yesterday and today and Brandon is working on the styling and I'm mentally exhausted at this point because I've been thinking about joining a join table with a resource for the past 3 hours, but things are slowly - very slowly - coming together